Fix linux max symlink depth problem
Macos finder will resolve a symlink when changing directory, like a
shortcut on windows. Linux file managers don't so you end up with a
massive chain of symbolic links. There is a limit to how many symlinks you
can chain together in one file path, so it eventually breaks.

Dolphin is capable of making and using .desktop files that behave like
windows shortcuts. It's the only common FM that I tried that could do this.
I tried: nautilus, thunar, nemo and pcmanfm.

Since dolphin is the only one that works, we unfortunately have to force
the player to install and use it.